requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
timm>=0.9.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
Pillow>=9.0.0
audiomentations>=0.30.0
